requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
python-dotenv
jinja2
aiofiles
google-generativeai
starlette